moved PhaseBadge into it's own component
