+++ src/widgets/PageHeader.tsx
@@ -3,6 +3,7 @@ import { jsx, css } from "@emotion/core";
 import React from "react";
 import { Link } from "@reach/router";
 import CdsLogo from "../CdsLogo";
+import PhaseBadge from "./PhaseBadge"
 
 const navStyle = css`
 color: white;
@@ -31,27 +32,11 @@ h1 {
 }
 `;
 
-const phaseBadge = css`
-  background: #f90277;
-  font-size: 1.2rem;
-  border-radius: 0.5rem;
-  color: white;
-  font-weight: 700;
-  padding: 0.2rem 0.8rem;
-  margin-left: 1rem;
-`
-
 const titleContainer = css`
   display: flex;
   margin-bottom: 1rem;
   align-items: center;
 `
-
-const PhaseBadge = () => {
-    return (
-      <span css={phaseBadge}>Alpha</span>
-    );
-  };
 
 const PageHeader = () => {
   return (
